requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
biopython=1.73=py36h7b6447c_0
blas=1.0=openblas
blast=2.9.0=pl526h979a64d_3
bzip2=1.0.8=h7b6447c_0
ca-certificates=2019.5.15=0
certifi=2019.6.16=py36_0
curl=7.64.1=hbc83047_0
cycler=0.10.0=py36_0
dbus=1.13.6=h746ee38_0
decorator=4.4.0=py36_1
entrez-direct=11.0=pl526_0
expat=2.2.6=he6710b0_0
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
glib=2.56.2=hd408876_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
hdbscan=0.8.22=py36hd352d35_1
hisat2=2.1.0=py36h2d50403_1
icu=58.2=h9c2bf20_1
joblib=0.13.2=py_0
jpeg=9b=h024ee3a_2
kiwisolver=1.1.0=py36he6710b0_0
krb5=1.16.1=h173b8e3_7
libblas=3.8.0=10_openblas
libcblas=3.8.0=10_openblas
libcurl=7.64.1=h20c2e04_0
libdeflate=1.0=h14c3975_1
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
liblapack=3.8.0=10_openblas
libopenblas=0.3.6=h6e990d7_4
libpng=1.6.37=hbc83047_0
libssh2=1.8.2=h1ba5d50_0
libstdcxx-ng=9.1.0=hdf63c60_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h1bed415_1
libxml2=2.9.9=hea5a465_1
llvmlite=0.29.0=py36hfd453ef_1
matplotlib=3.1.0=py36h5429711_0
mcl=14.137=pl526h470a237_4
ncurses=6.1=he6710b0_1
networkx=2.3=py_0
numba=0.44.1=py36hb3f55d8_0
numpy=1.16.4=py36h99e49ec_0
numpy-base=1.16.4=py36h2f8d375_0
openblas=0.3.6=h6e990d7_4
openssl=1.1.1c=h7b6447c_1
pandas=0.24.2=py36he6710b0_0
patsy=0.5.1=py36_0
pcre=8.43=he6710b0_0
perl=5.26.2=h14c3975_0
perl-app-cpanminus=1.7044=pl526_1
perl-archive-tar=2.32=pl526_0
perl-base=2.23=pl526_1
perl-business-isbn=3.004=pl526_0
perl-business-isbn-data=20140910.003=pl526_0
perl-carp=1.38=pl526_3
perl-common-sense=3.74=pl526_2
perl-compress-raw-bzip2=2.086=pl526hf484d3e_0
perl-compress-raw-zlib=2.086=pl526h6bb024c_1
perl-constant=1.33=pl526_1
perl-data-dumper=2.173=pl526_0
perl-digest-hmac=1.03=pl526_3
perl-digest-md5=2.55=pl526_0
perl-encode=2.88=pl526_1
perl-encode-locale=1.05=pl526_6
perl-exporter=5.72=pl526_1
perl-exporter-tiny=1.002001=pl526_0
perl-extutils-makemaker=7.36=pl526_1
perl-file-listing=6.04=pl526_1
perl-file-path=2.16=pl526_0
perl-file-temp=0.2304=pl526_2
perl-html-parser=3.72=pl526h6bb024c_5
perl-html-tagset=3.20=pl526_3
perl-html-tree=5.07=pl526_1
perl-http-cookies=6.04=pl526_0
perl-http-daemon=6.01=pl526_1
perl-http-date=6.02=pl526_3
perl-http-message=6.18=pl526_0
perl-http-negotiate=6.01=pl526_3
perl-io-compress=2.086=pl526hf484d3e_0
perl-io-html=1.001=pl526_2
perl-io-socket-ssl=2.066=pl526_0
perl-io-zlib=1.10=pl526_2
perl-json=4.02=pl526_0
perl-json-xs=2.34=pl526h6bb024c_3
perl-libwww-perl=6.39=pl526_0
perl-list-moreutils=0.428=pl526_1
perl-list-moreutils-xs=0.428=pl526_0
perl-lwp-mediatypes=6.04=pl526_0
perl-lwp-protocol-https=6.07=pl526_4
perl-mime-base64=3.15=pl526_1
perl-mozilla-ca=20180117=pl526_1
perl-net-http=6.19=pl526_0
perl-net-ssleay=1.88=pl526h90d6eec_0
perl-ntlm=1.09=pl526_4
perl-parent=0.236=pl526_1
perl-pathtools=3.75=pl526h14c3975_1
perl-scalar-list-utils=1.50=pl526h14c3975_0
perl-socket=2.027=pl526_1
perl-storable=3.15=pl526h14c3975_0
perl-test-requiresinternet=0.05=pl526_0
perl-time-local=1.28=pl526_1
perl-try-tiny=0.30=pl526_1
perl-types-serialiser=1.0=pl526_2
perl-uri=1.76=pl526_0
perl-www-robotrules=6.02=pl526_3
perl-xml-namespacesupport=1.12=pl526_0
perl-xml-parser=2.44=pl526h4e0c4b3_7
perl-xml-sax=1.02=pl526_0
perl-xml-sax-base=1.09=pl526_0
perl-xml-sax-expat=0.51=pl526_3
perl-xml-simple=2.25=pl526_1
perl-xsloader=0.24=pl526_0
pip=19.1.1=py36_0
pyparsing=2.4.0=py_0
pyqt=5.9.2=py36h05f1152_2
python=3.6.8=h0371630_0
python-dateutil=2.8.0=py36_0
pytz=2019.1=py_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
samtools=1.9=h8571acd_11
scikit-learn=0.21.2=py36hcdab131_1
scipy=1.3.0=py36h921218d_0
seaborn=0.9.0=py36_0
setuptools=41.0.1=py36_0
sip=4.19.8=py36hf484d3e_0
six=1.12.0=py36_0
sqlite=3.28.0=h7b6447c_0
statsmodels=0.10.0=py36hdd07704_0
tk=8.6.8=hbc83047_0
tornado=6.0.3=py36h7b6447c_0
umap-learn=0.3.9=py36_0
wheel=0.33.4=py36_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
